grammar: replace "which" by "that" where it is a defining or restrictive pronoun.
Although not mandatory, this distiction is recommended by many
authorities (eg. Strunk and White, APA Publication Manual, and
Fowler).  The copyright for Fowler and Fowler's "The King's English"
has expired so it can be found online (it's a bit of a tough slog but
I find it entertaining):

<http://www.bartleby.com/116/205.html>

Several other improvements were discovered at the same time.