pluto: recalculate default SPD priority using calculate_sa_prio(c)
Once we instantiate and fill in the details, we need to have a longer
prefix match priority for the instance compared to the template, so
packets hit the instance SPD and not the template SPD rule.

The replaced priority creation also could not make a difference between
static conns, OE conns and OE anon conns. This is now differentiated so
that SPDs (trap or tunnel!) for static tunnels always wins over OE.

Note for bare shunts where we havent determined a connection yet, we use
prio 0 (the highest). This is not a change from previous behaviour.