ikev2: map DH=&ike_alg_dh_none into .st_pfs_group=NULL when accepting response
In IKEv2, a proposal that contains neither INTEG nor DH transforms can
be valid - in both cases it is interpreted as proposing 'none'.  When
this happen, common code expects INTEG = &ike_alg_integ_none and DH =
NULL.

This would all be ok except that there's a bug in the proposal code
when --impair allow-null-none [sic].  The result is a core dump
(INTEG=NULL, oops), and the simplest fix ends up also 'fixing' DH (as
in it will be set to &ike_alg_dh_none instead of NULL).

This patch prevents DH = &ike_alg_dh_none getting into common code.