whack: Ensure all status and list commands return no error code
So even if the status or list will report an error from whack_log()
this will not be copied into the return code for whack, since the
status or list command itself succeeds fine.

Of course, failing to use the whack socket for status/list commands
still gives a non-zero error code.