logging: libreswan_loglog() -> libreswan_log_rc()
The only difference between libreswan_log() and above is that the
above takes an extra RC parameter.

I suspect the name *loglog() dates back to when it and *log() were
different - the former would send output to both syslog and whack
while the latter would only sending output to syslog.  That's no
longer true; and reflecting this the short form loglog() should
probably be renamed to something.