pluto: add big 'exiting_pluto' flag, check in fetch and helper threads
From the comment:

The global EXITING_PLUTO is there as a hint to long running threads
that they should also shutdown (it should be tested in the thread's
main and some inner loops).  Just note that, on its own, it isn't
sufficient.  Any long running threads will also need a gentle nudge
(so that they loop around and detect the need to quit) and then a join
to confirm that they have exited.

Also avoid pthread_cancel() which can crash.