testing: fixup ikev2-initial-contact-send
The test needed generic updating. The test also checked for
proper sending of Initial Contact via a debug line that could
miss an actual difference in number of payloads send if a future
error is introduced. Instead, the test now checks the log for
the actual number of payloads sent and received.