building: sort all wildcarded object files for build reproducibility
Without this sort, the linker may assemble the final object files in
the order in which their source files are found in the filesystem (or
however make chooses to interpret the $(wildcard ...) function).

By sorting the results lexicographically, we aim to assemble
byte-identical output files, regardless of the filesystem order or
other possible non-determinism introduced by make.

This resolves https://github.com/libreswan/libreswan/pull/217

Signed-off-by: Paul Wouters <pwouters@redhat.com>