libswan/asn1.c: another implementation of is_printablestring()
This one is probably faster than the original (not measured).

It is probably faster and more straight-forward than
  62a04f2022e4b450d87b0f9eacb1ea7840c2701f
The new one doesn't need to scribble on its input.

The character set is a little harder to change, but the standard
dictating it is very unlikely to change.  (I generated the table with
a program but the chance of ever having to run it again is zero, so
I'm throwing it away.)