testing: set test domain disks to cache=unsafe
This turns off KVM's code trying to flush virtual disk writes to
the physical (host) disk.  Which, in turn, should improve performance.
(as measured by http://testing.libreswan.org/results/ which doesn't
use tmpfs, if this doesn't show improvement it can be switched off).

The down side is that a host crash will likely leave the test
domain's disk corrupt.  That should be ok as the test domain
disks are re-cloned every time libreswan is installed.