IKEv2 TS: compute the 'narrow port' and then assign a score
Where the 'narrow port' is either a single port or the entire port
range.

Should ensures that only things that 'struct end' can describe get a
positive score.  For instance, it will not try to narrow end.port=0
(aka 0-65535) down to TS.ports=1-65534 because that can't be described
by 'struct end'.  While it could narrow down to '1', there's no test
showing a need for this.