ikev2: always log successful state transitions
Contrary to the comment "tell whack and log of progress, if we are
actually advancing", the lack of a state change does not imply a lack
of progress.  For instance, when the initiator gets back
INVALID_KE_PAYLOAD, it does progress but does not change state.

It should probably instead log the state transition's story, but
that is, well another story.