clarify interface to several event-deletion functions
Each deals with a pointer to a pointer to the event.

If the pointer to the event is NULL, the function does nothing.  This
is considered normal and boring: no logging is performed.  So callers
need not check for this case.  This was true for some functions but
not all functions and it was not documented

The pointer to the event is set to NULL by the deletion routine.
This eliminates a dangling pointer.  Previously this was usually true
but not always and was not documented.

If the event pointer is not NULL, it must point to a known event.
On pain of assertion failure.

Simplify callers by exploiting these interface changes.

Simplify find_ifaces().

Inline unlink_pluto_event().

const added; spelling fixed.