pluto: Update the logic of NAT-T keepalives
The logic is based on the assumptions that:

1) Running get_sa_info() on IPsec SA's is (too) expensive to do every 20s
2) Looking up IKE SA's for IPsec SA's is expensive

So we do not know if an IPsec SA is idle for the keepalive check.

For IKEv2, don't bother with keepalives for IPsec SA, as every IPsec SA has
an IKEv2 IKE SA. Simply check the IKE SA and send keepalives there.

For IKEv1, IPsec SA's can have no IKE SA. So we cannot skip the keepalive
checks for the IPsec SA  and only use the IKEv1 IKE SA. So since we are
sending the keepalives on the IPsec SA state, we might as well not bother
sending them for the IKEv1 IKE SA.