EMPTY_CHUNK: add const qualifier to type of EMPTY_CHUNK
Since a C compound literal is an object, it can be modified.
EMPTY_CHUNK expands into a compound literal.
In theory, every use of EMPTY_CHUNK creates a distinct object.
Adding "const" to the type signals the compiler and the
programmer that these objects cannot and must not be modified.