testing: fixup RSA-PSS test cases
These are not using a CA, but depend on having preloaded certificates.
However, in ipsec.conf.common we had removed this recently, as it was
interfering with tests of CERT payloads over IKE. So now these tests
load the certs explicitely on the libreswan endpoint. Also some minor
output updates for strongswan