v2 msgids: replace SMF2_SEND with enum message_role .send
So state m/c can specify that the outgoing message is a request,
response, or no-message (default).

While it is likely possible to figure this out using just the
SMF2_SEND flag and the incomming message (or lack there of) it is
easier and clearer to spell it out.

Troubling is the number of state transitions (for instance when
resonding to a notification request) that don't specify any
outgoing response - look for record 'n' send 'n' delete.