demux: add msg_digest .fake_dne, set when creating a fake_md
This way v2_msg_role() can treat a NULL MD and a fake MD the same
and return NO_MESSAGE - which is true because the fake_md gets
created when there isn't a message.

Note that this is different to is_msg_{request,response}() which
just look at the flags in the header.