v2 msgids: hack delete request so that the the current Message ID is saved in the IKE SA's .st_msgid
It was saving the next Message ID (i.e., intended for the next
message) in the CHILD SA's .st_msgid.  Kind of futile given that the
immediately after all this the CHILD SA was deleted.  Things were
hanging together because the state m/c, on finding no CHILD SA, was
routing the the message to the IKE SA anyway.