pervasive: concatenate ASN.1 algorithm blobs with their length field
Each use of an algo blob prepended the blob with the length field.
This change does so in the blob table rather than requiring each use
to do it.  This simplifies the code.  In the process, the last remaining
VLA (that Hugh was aware of) has been eliminated.

Added context to a few messages.