proposals: add a new parser that allows multiple algorithms of the same type per proposal, enable for IKEv2
For instance:

   ike=aes_gcm+aes+ccm

Currently the parser doesn't suppress duplicates such as
aes_gcm+aes_gcm and aes_gcm,aes_gcm.  The algparse tests will
fail because of this.

Hand testing sugests it works.

(There are still hard-coded limit in the IKEv2 proposal code that may
need bumping up).

IKEv1 is left as an exercise for the developer.