ikev2: don't segv when, while expiring a CHILD SA, the parent is missing
Seems to happen because the CHILD SA failed to establish, and started
replacing simultaneous to, or perhaps triggering, the IKE SA to also
trying to replace.  Or is that the other way around?

Would a sequence like:

- generate create events for all existing child connections
- wipe out the entire family

be safer?  Currently this, and children all seem to overlap leading to
fights like the above.