testing: WIP added certoe-13-poc-del-slash24 and certoe-14-poc-del-slash32
The first shows a /32 in a /24 re-establishment. It is almost right,
but for some reason delays showing OE state on road.

The seconds shows a /32 in a /32 re-establishment. It shows a bug in
where the groupinstance is /32 and the OE tunnel itself (also /32)
is removed, no groupinstance policy is re-instated.