IKEv1: Remove a bogus "duplicate Delete" check which prevented immediate re-initialize
This check dates back to openswan-1.0.0 and was apparently to prevent
spurious duplicate deletes from Windows. Possibly, a misunderstood case
of Windows sending an IPsec delete followed by an IKE delete. It checks
if we are within 60 seconds of an EVENT_SA_REPLACE event, and skips
the replace action.

It caused the IPsec SA delete to be ignored without replace, causing
an outage for 60 seconds or until the other end initiates.