testing: added basic-pluto-01-failtest
This test does not use iptables but relies on the new helper
xfrmtest.sh to show errors when one side receives an unexpected
plaintext packet.

Note this does not cover yet the case where both ends make a mistake
and send/accept clearly. For that, we really need to tcpdump all
traffic, then re-read and count the ESP packets :/