connections: merge secret and cert code so things get done less often
As part of trying to pre-load a connection's private key (aka secret),
the code was searching for and then loading the certificate only to
then throw it all away.

This patch merges that code with the connection's end code which also
searches for and then loads the certificate.

The change also:

- remembers to free the cert when fips rejects it

- rejects the connection when a cert's low-level details can't be
  extracted