connections: allocate a connection and then use extract_connection() to fill it in
Replace code populating the connection with mash up of shared pointers
(shared with static memory, shared with the whack message, shared with
a certificate, ...) and allocated data and then trying to unshare
things

Instead always allocate the connection and then populate it.  When
things go wrong just call discard_connection().