kernel: sort Linux's IPv6 interfaces before adding them to the "interfaces" list
The order of /proc/net/if_inet6 can't be trusted.

The order is: protocol {loopback,normal,any} ifacename addr port;
which is a little over the top.  Should IPv4 also be sorted?

Note that the "add" code prepends each interface to the "interfaces"
list causing the final order to be reversed!?!?