IKEv2: expire_ike_because_child_not_used() should use was_eroute_idle()
It was using manual calls to get_sa_info() to determine idleness, but we
have the function was_eroute_idle() for that.

Also leave a note that "idleness" is oddly based on the rekeymargin= value.

No need to address this now, soon we should install softidle timers into the
kernel so the kernel will tell us when an IPsec SA is idle, instead of having
IKE keep asking the kernel and then checking it manually.