testing: re-enegineer ikev2-55-ipseckey-02
Due to a bug, likely in the dns handling on east, this test failed.

But this test's primary purpose was to test two keys in DNS where road
can change key at any time and still work. So this test is redone to
focus on the road test. This new test failure is copied as a new test
case ikev2-55-ipseckey-09