ikev2: replace processed_retransmit() et.al. with duplicate_{request,response}()
The fields .v2_msgid_windows and .v2_msgid_wip track initiator and
responder Message IDs independently.  By assuming this is working
correctly, the need to look at the current state_kind and other magic
is eliminated (and if something goes wrong, the first suspect should
be a missing Message ID update).

The old code was using .st_msgid to simultaneously track both the
current request and response Message ID.

In addition the hack setting the responder's IKE SPIs from the
IKE_SA_INIT response is moved to after the duplicate/busy checks; the
redundant duplicate checks in the code looking for the state are
eliminated; and an O(#STATES) search is removed.