id: hack up jam_id() and jam_id_escaped()
Implement using idtoa() and escape_metachar() so that the wierd
sequence of custom sanitization is preserved. idtoa() sanitizes
using '?' and escape_metachar() sanitizes using '\NNN'.   However,
the later is only ever used to sanitize idtoa() output so !isprint()
get sanitized as '?' and not '\NNN'.  Which, of course, is different
to sanitize_string().  Ulgh!