testing: extend dist_certs.py to create a bunch of weird certificates
Various combinations of Basic Constraints (BC), Key Usage (KU),
Extended Key Usage (EKU) and Critical Flags set or omited on the
certificate. Certificates that are expected to fail validation
according to the KU/EKU/CF/BC criteria have a "-bad" suffix.

Unfortunately, some of the options we want to test are refused by
openssl for certificate generation. The code is left in to try,
but the certificate name is commented out so the attempted
generation doesn't cause the script to fail.