documentation: Fix spelling and grammar
therefor -> therefore
transfered -> transferred
acceptible -> acceptable

This also involves some clean up of "allow to", the use of which
typically presents some amount of confusion about who is doing the
allowing, and who is being allowed.  Given that the options in
question are all called accept-*, it makes more sense to describe them
as "accepting" anyway.

Signed-off-by: Paul Wouters <pwouters@redhat.com>