logging: replace log_to_log() with plog_global() and plog_st() et.al.
Since, apparently, at one point the old plog() function never sent
messages to whack, use that as the prefix.

Since the context (state, connection, from), or lack there of, is
explicit these functions are not reliant on the cur_* globals.

As a gentle nudge towards using the context sensitive functions
(plog_st()), the context free function is given the longest name.

Repainting the wagon.