crypto: replace MD with MDP as parameter to crypto_completed_cb(), yuck
In IKEv2, the only place that *MDP is "stolen" is in
complete_v2_state_transition() when handling STF_SUSPEND.  Hence,
there's no good reason for passing in a pointer to MD a.k.a. MDP.

However, in IKEv2, this isn't the case - the code base is still
littered with suspend_md() calls.  Since that isn't likely to change,
MD is replaced with MDP making the interface IKEv1 code base friendly.
Yuck.