ikev2: when adding a connection, use the proposal parser to generate default proposals
Instead of leaving the proposals NULL and then, in the middle end,
using defaults from static structures.

As a consequence:

- when updating the IKEv2 default proposals, only the proposal parser
  needs to be modified

- the code in ikev2_spdb_struct.c dealing with NULL proposals is dead

- eliminates an obscure bug in the code generating duplcate MSDN
  proposals from the static structures (it missed one proposal, the
  code is no longer used, and the non-static code path works)

Because the IKEv2 proposal parser can exactly describe the default
proposal suite there is no externally visible change (well, other
than fixing that obscure bug).  This isn't true of IKEv1 (but that
doesn't mean a similar change to IKEv1 isn't a good idea).