connections: drop host_pair's host_port_specific field and port parameters
Since the field .host_port_specific is never set (i.e., always false)
the port check in:

  sameaddr(&p->him.addr, myaddr) &&
  (!p->him.host_port_specific || p->him.host_port == hisport)

is always true.

Drop the field and parameters.  While we're at it, replace {me,him}
with local and remote endpoints and use str_endpoint().

Add XXX: comment to the two places (not even in hostpair.[hc]) where
the port is used.