chunks: replace CHUNK0() with THING_AS_CHUNK() - generates a warning when THING is constant
Convert writeable THING to a writeable CHUNK.  When compiled with GCC
(at least) and THING is read-only, a warning will be generated.

This works because GCC doesn't like implictly converting a 'const'
&THING actual parameter to the non-const 'void*' formal parameter.
Using an explicit cast (such as in a static initializer) suppresses
this warning.

For a read-only CHUNK like object, see THING_AS_SHUNK().