pluto: redo audit logging. Centralize is better.
This did uncover limitations in the state machine where esepecially
for IKEv1, failures do not cause a state change where we can hook
in the failure logging.

Additionally, in IKEv2, a child SA failure leads to a parent SA
failure, where instead we should complete and keep the IKE SA and
only fail the child sa.