state: initialize .st_local{addr,port} using .st_interface's .local_endpoint
Since it is the interface's local_endpoint that will be used when
sending the message.

Was starting with the connection's .this.end (orientation means it can
be different to the interface's local_endpoint), then later realising
it didn't match the interface and "floating" it across.

(update_ike_endpoints() was already syncing .st_local{addr,port} with
.st_interface).