pluto: server.c: create_socket(): simplify port handling
The previous version overwrote ifp->addr's port, used ifp->addr, and
then rewrote it to pluto_port, but only if the bind call succeeded.

That's confusing and posibly buggy on the bind failure path.

This commit leaves ifp->addr alone and changes the port on a copy.

I don't know if ifp->addr's port matters.  The name "addr" is
unpleasant to grep for so I didn't.