x509 logging: in jam_id_escaped()/str_id() sanitize using jam_*_bytes()
That is, sanitize the data as it gets added to the buffer, and not as
a second pass.

This changes (arguably fixes) the output of ID shell variables:

- unprintables are emitted as \ooo (instead of '?') (so scripts can
  recover the actual characters)

- meta-characters are emitted as \ooo (for instance, \t is emitted as
  \011 instead of \0311, don't ask)

Follows on from 7a0f392be772bec73a05d67f1ca5d9d338ec8fa8 which added
jam_id(jambuf_t*,id,jam_bytes_fn).

Also follow on from 60956ab1eb15e13ab69aa8e010187ea653cdf4ef which
eliminated most of the idtoa() calls.  What to do with the remaining
idtoa() is an open question - currently it still emits '?' for
unprintable.