pluto: new option require-id-on-certificate=yes|no
When using X.509 certificates, this option can be used to accept certificates
that violate the rules of RFC 4945 Section 3.1 by not having their IKE ID
listed as a subjectAltName (SAN) on their certificate. The default (yes)
is to not accept these certificates as it enables an attack where a compromised
host can use its valid certificate and some other hosts' peer ID to pretend
to be that other host.

This commit also fixes decode_peer_id_counted() called via decode_peer_id()
that would mistakenly skip the peer ID check on the responder, resulting in
IKE peer ID's not specified on the certificate to be accepted on responders.