pluto: pass --config option from pluto to forked addconn
Otherwise the addconn that is started by pluto will use the
compiled in default /etc/ipsec.conf and not pick up the same
config as pluto was started with.

Also, store this properly for displaying in ipsec status, as
the libipsecconf parser inside pluto isn't updated with this
non-standard config file location.

This is rhbz#1760633