logging: only dup a whackfd when needed, typically when storing it in a heap object
For instance, when copying the "global" whackfd into pending; and
copying pending's whackfd into the state.  Conversely, don't dup the
wackfd when it is being passed as a parameter on the stack.

In first_pending(), explicitly release the child's old whackfd before
storing a duped new value.

In release_pending_whacks(), when ST is a CHILD SA and it's IKE SA has
a matching whackfd, release all the IKE SA's pending children (stops
whack being left haing when the child fails; the code was trying to
release the CHILD SA's pending children ...).

In flush_incomplete_child(), explictly release the child (stops extra
log messages sneaking through to whack).

This replaces "ownership transfered".