ikev2: replace state processor's ST with CHILD
Instead of the noncommital 'struct state *st' parameter (is it an IKE
SA or is it a CHILD SA) pass in an explict CHILD when known.  In the
processors pexpect a NULL or non-NULL CHILD.

Add note that, perhaps, the CHILD parameter should be dropped -
instead deal with children inside an IKE processor.