testing x509: add fake EC certs, kind of, sort of
The intent is to add EC versions of the mainca certs so it
is possible to screw around with DNs that match wrong
certs.

The bug, also in curveca code, that all the certs are
self signed isn't fixed.  The second bug that all certs
had the same key is fixed.