ikev2: set the REPLACE timeout to 2*retransmit timeout when sending IKE_AUTH request
Both retransmit and replace (for half-open IKE) had the same
value 60s leading to either being delivered first.

If the replace is delivered first, whack was left hanging (the replace
keeps the socket open, even when the retransmit is trying to close
it).  Slowing down the replace in once sense avoids the problem but in
another sense fixes it as the retransmit timeout should take priority.