demux: rename msg_digest's .from_state to .v1_from_state; make it IKEv1 only
In IKEv2, .from_state was a redundant copy of md->svm->state (which in
turn is, or at least should be, a redundant copy of
st->st_state->kind).

Tweak process_v2_child_ix() to use svm->state making it clearer that
its behaviour is determined by a field in the state transition struct.

Can the same be said of IKEv1?