pluto: ISAKMP_SA_established() should support reconnecting authnull clients
This function releases old connections (on a different or same IP) of the
same IKE credentials. If the new connection was NULL authenticated, it
would never replace as the identities cannot be verified. This caused a
problem for a crashing client that reconnects while the remote peer still
has the connection up. The remote peer would reject the new attempt with
an "eroute already in use" error.

This condition is now detected and allowed. An additional check had to be
made to ensure two NULL auth connections from different IPs do not replace
each other based on "the same IKE authentication".

This fixes newoe-27-replace-sa-authnull-authnull