pluto: update calculate_sa_prio() to allow OE shunt to override priority.
The OE shunt is called with the OE group template, which can be /24 for the
remote group. But we need to calculate the SA priority for the shunt as if
it was a /32 (or /128) so that the %pass shunt will have a higher priority
(lower number) than the OE group %trap policy it was triggered from.