IKEv2: fix PPK code that wasn't following RFC draft
the situation was the following:
On responder side the peer (responder) didn't find a suitable PPK
but he was configured to use PPK (ppk=insist). This happens in IKE_AUTH exchange.
Instead of sending AUTHENTICATION_FAILED back, the responder silently failed.

See bottom of page 9 of draft-ietf-ipsecme-qr-ikev2-11.

Update the console output in one affected test.

Signed-off-by: Paul Wouters <pwouters@redhat.com>