IKEv2: Verify (not ignore) expected TSi/TSr payloads for IPsec rekeys.
The RFC states they MUST be the same for IPsec rekey as they are for the
original exitsing IPsec SA. So for rekeying, the traffic selectors were
just copied from the existing IPsec being rekeyed and ignored the actual
TSi/TSr payloads during IPsec rekey.

It now parses these and returns INVALID_SYNTAX if it fails to parse. This
happens on libreswan 3.30 due to a bug in the IPsec rekey code on that version.

It also verifies the traffic selector that was negotiated for the existing
IPsec SA is still an entry on the received TSi/TSr list. Other entries on
the list are ignored.