show: poke the spacer dragon
Pass a bool between show_shunt_status() and show_states() so
that they can indicate if/when a spacer (blank line) is needed
before the next output.

Rather than add the bool as an extra parameter, embed it in the
new struct show (along with the whackfd) and pass that.  Hopefully
makes further changes less tedious.