seamless nss cert rotation for existing tunnels
Libreswan does not support seamless cert rotation. Therefore, when
a new cert is issued and inserted into NSS DB, a tunnel should be
torn down and set up again or pluto needs to be restarted in order
to pick up the new cert. This causes temporary disruption for an
existing tunnel. This patch enables a new feature that allows
rotation of cert without bringing down any active tunnel by
replacing an old cert with a new one in pluto.