pluto: add FIPS validation for EC keys for ECDSA
This change closes 318.

Prior to this change, there was no FIPS check for elliptic curve
cryptography (ECC) key curves used for ECDSA. So what happened was that
FIPS verification for RSA keys would run when EC keys were used which
resulted in failures.

The FIPS check logic for an EC curve in this change is identical to what
`certutil` does. That is, the logic uses Network Security Services (NSS)
to determine what curves are allowed in FIPS mode. NSS, meanwhile,
relies on the operating system security policy to know what is allowed.

An alternative implementation to check FIPS approval for an EC curve is
to explicitly check against the named curves listed in Appendix D in
https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.186-4.pdf I decided
against this approach because it would require editing the code if the
approved set of curves changes in the future. It is better instead to let
the OS security policy that configures NSS to dynamically specify what is
allowed.