msgid: add last_contact
Track the IKE initiator and responder last contact (time a message
was received) separately.

Getting a response is a good sign that both the outbound and
inbound paths are still working.

Getting a request just indicates that the incomming path
is working (for the moment ignore how a new request indicates
that the previous request must have succeeded; suspect mobike
breaks this).